Improve resize listener with the `loadedmetadata` event in `DailyVideo`
The current `onResize` callback may not fire on initial load. Since it
relies on React’s render timing (using a combination of `useEffect` and
`requestAnimationFrame`) instead of events on the video, the video
stream may not be loaded in time for the initial `handleResize()` call.

React has built-in support for `onResize` and `onLoadedMetadata` on
`video` elements. This commit ditches the custom resize listener defined
in `useEffect` in favor of these supported properties.