simplify custom participant ids
Instead of using recoil's transaction observer, we can essentially calculate custom ids (for custom filter and/or sort) with a combination of useCallback and useMemo. The useMemo only re-calculates when one of the input parameters (filter, sort, or the standard list of ids) changed. This is less optimal compared to string-based filter and/or sort instances of useParticipantIds: the returned array's reference will change, when e.g. participants join or leave while the returned ids do not, but this is still a lot better and a fairly small trade-off, given the flexibility of custom filter and sort.